Add procedure pointer (cfun) to basetree_t objects (WIP)
Passing of "user compare_function" via all calls to tree routines
is obsolete. All routines should use a procedure pointer associated
with the user function.
Code transformation status:
the function is in the structure and associated, but old interface
is still used. I want this commit as a back-up before modifying
all procedures to use this pointer associated routine.